Split data models from view models (Models/ and Views/)
- Create Models/ with pure data types: InspectionResult, AssemblyAudit, RidPackageReference
  Zero Markout references in any model file
- Create Views/ with Markout presentation: InspectionResultView, AssemblyAuditView,
  AssemblyAuditReport, AssemblyDependenciesView, FlatDependency, RidPackageReferenceView
- OutputFormatter pivots: JSON serializes data models, Markout serializes view models
- Move FlatDependency to Views/ (only created by and used in view serialization)
- All 375 tests pass (253 app + 122 services)
