+++ src/dotnet-inspect/Commands/AssemblyCommand.cs
@@ -1,3 +1,4 @@
+using DotnetInspector.Models;
 using DotnetInspector.Inspectors;
 using DotnetInspector.Metadata;
 using DotnetInspector.Options;
