+++ src/dotnet-inspect/Commands/PackageCommand.cs
@@ -1,3 +1,4 @@
+using DotnetInspector.Models;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using DotnetInspector.Inspectors;
